fixed staff

+++ corebanking/src/main/java/com/corebanking/repository/AnnouncementsRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.corebanking.entity.Announcements;
 
-public interface AnnouncementsRepository 
+public interface AnnouncementsRepository
 
-          extends JpaRepository<Announcements, Long> {
+          extends JpaRepository<Announcements, Integer> {
 }
